# packages/contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Smart Travel Assistant — Public API
  version: 1.0.0
  summary: The only HTTP surface a browser is allowed to talk to.
  description: |
    Public trust boundary for the Smart Travel Assistant (module 02). The browser reaches
    `web` and `api` and nothing else: internal services are not routable from outside the
    Docker network, so every capability a user has is expressed here.

    Rules that hold for every operation in this document:

    - Authentication is an OIDC access token issued by Keycloak, verified against JWKS
      (issuer, audience, signature, expiry, not-before). `/health/*` is the only exception.
    - Ownership is resolved from the token, never from a body field. A request that names
      another user's resource gets 404, not 403, so that ids cannot be probed.
    - Times are ISO-8601 with an explicit offset. Coordinates are GeoJSON `[longitude, latitude]`.
    - Every mutating request accepts `Idempotency-Key`. Replaying a key with the same payload
      returns the original result; replaying it with a different payload is `IDEMPOTENCY_CONFLICT`.
    - Every response carries `meta.request_id`, `meta.contract_version` and `meta.generated_at`,
      and lists any dependency that degraded while producing it.
    - Anything safety-relevant carries its source, freshness and quality. When a provider or
      credential is missing the response says so explicitly; it never substitutes a plausible value.
...

...
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/trips:
    get:
      tags: [trips]
      operationId: listTrips
      summary: List the caller's trips
      description: |
        Every query is filtered by the owner resolved from the token, at the repository layer
        rather than in the handler, so an ownership check can never be forgotten. Soft-deleted
        trips are excluded unless `status=DELETED` is requested explicitly.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ContractVersion'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          required: false
          schema:
            $ref: '../jsonschema/common/enums.schema.json#/$defs/TripStatus'
      responses:
        '200':
          description: Trips owned by the caller, newest departure first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripListResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [trips]
      operationId: createTrip
      summary: Create a trip
      description: |
        Both endpoints must be confirmed locations. Departure time is validated against the
        configured backdating allowance and the travel modes against provider coverage for the
        region: a mode with no real data source is rejected here rather than silently assessed.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ContractVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
            examples:
              trip:
                $ref: '#/components/examples/CreateTripExample'
      responses:
        '201':
          description: Trip created at revision 1.
          headers:
            ETag:
              schema: { type: string }
              description: Weak ETag carrying the trip revision, e.g. `W/"1"`.
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '422':
          description: |
            The request is well-formed but cannot be served: for example a travel mode with no
            provider coverage in that region (`UNSUPPORTED_COVERAGE`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/trips/{trip_id}:
    parameters:
      - $ref: '#/components/parameters/TripId'
    get:
      tags: [trips]
      operationId: getTrip
      summary: Get one trip
      description: |
        Returns the trip with an `ETag` carrying its revision. Clients send that value back in
        `If-Match` when updating.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ContractVersion'
      responses:
        '200':
          description: The trip.
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [trips]
      operationId: updateTrip
      summary: Update a trip
      description: |
        Requires `If-Match` with the revision the client last read. A stale revision is rejected
        with 412 so that a second tab cannot overwrite a change it never saw. Changing anything
        that affects the journey invalidates the current assessment: the client must start a new
        one rather than keep showing the old result.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ContractVersion'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTripRequest'
      responses:
        '200':
          description: Updated trip at the next revision.
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
    delete:
      tags: [trips]
      operationId: deleteTrip
      summary: Soft-delete a trip
      description: |
        Marks the trip deleted immediately and schedules the purge of its child data. The
        response reports the deletion status so the UI can tell the user that removal is in
        progress rather than implying it already finished everywhere.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ContractVersion'
      responses:
        '200':
          description: Deletion accepted, with its current status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionStatusResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC access token issued by the project's Keycloak realm. Verified against the realm's
        JWKS: issuer, audience, signature, expiry and not-before are all checked, and an
        unverifiable token is rejected with 401 — never decoded unsigned as a fallback.
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: |
        Client-supplied correlation handle. The server generates one when absent and always
        echoes the value it used in `meta.request_id`.
      schema:
        $ref: '../jsonschema/common/primitives.schema.json#/$defs/Uuid'
    ContractVersion:
      name: X-Contract-Version
      in: header
      required: false
      description: Major contract version the client was built against. Currently `1`.
      schema:
        type: string
        pattern: '^[0-9]+$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Client-generated key that makes a retry safe. The same key with the same payload returns
        the original result; the same key with a different payload is rejected with
        `IDEMPOTENCY_CONFLICT`. Keys expire after the configured retention window.
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    TripId:
      name: trip_id
      in: path
      required: true
      schema:
        $ref: '../jsonschema/common/primitives.schema.json#/$defs/Uuid'
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from a previous page. Not to be constructed by the client.
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    ErrorResponse:
      $ref: '../jsonschema/common/envelope.schema.json#/$defs/ErrorResponse'
...
    CreateTripRequest:
      type: object
      additionalProperties: false
      required: [origin, destination, departure_time, timezone, travel_modes]
      properties:
        title:
          type: [string, 'null']
          maxLength: 256
        origin:
          $ref: '#/components/schemas/LocationRef'
        destination:
          $ref: '#/components/schemas/LocationRef'
        departure_time:
          $ref: '../jsonschema/common/primitives.schema.json#/$defs/Timestamp'
        return_time:
          $ref: '../jsonschema/common/primitives.schema.json#/$defs/NullableTimestamp'
        timezone:
          $ref: '../jsonschema/common/primitives.schema.json#/$defs/Timezone'
        travel_modes:
          type: array
          minItems: 1
          maxItems: 7
          uniqueItems: true
          items:
            $ref: '../jsonschema/common/enums.schema.json#/$defs/TravelMode'
        preferences:
          $ref: '#/components/schemas/TravelPreference'
...
    TripResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/Trip'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    TripListResponse:
      type: object
      required: [data, meta, page]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Trip'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        page:
          $ref: '#/components/schemas/PageMeta'
...
  responses:
    ValidationError:
      description: The request failed validation. `field_errors` names each rejected path.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              $ref: '#/components/examples/ValidationErrorExample'
    Unauthorized:
      description: No token, or a token that failed verification.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    IdempotencyConflict:
      description: The idempotency key was reused with a different payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  examples:
...
    CreateTripExample:
      summary: Creating a trip from two confirmed locations
      value:
        title: Bangkok to Chiang Mai
        origin:
          place_id: '1609350'
          display_name: Bangkok, Thailand
          coordinates:
            type: Point
            coordinates: [100.5018, 13.7563]
          country_code: TH
          admin1: Bangkok
          timezone: Asia/Bangkok
          provider: open_meteo
          confirmed_by_user: true
        destination:
          place_id: '1153671'
          display_name: Chiang Mai, Thailand
          coordinates:
            type: Point
            coordinates: [98.9817, 18.7883]
          country_code: TH
          admin1: Chiang Mai
          timezone: Asia/Bangkok
          provider: open_meteo
          confirmed_by_user: true
        departure_time: '2026-09-20T09:30:00+07:00'
        return_time: null
        timezone: Asia/Bangkok
        travel_modes: [TRAIN]
        preferences:
          prefer_safer_route: true
          prefer_lower_cost: false
          prefer_lower_emissions: false
          max_extra_duration_minutes: 90
          avoid_tolls: false
          accessibility: []
...